[processmodules] reduced code required to add new process modules for robots
