+++ pycram.rosinstall
@@ -23,3 +23,7 @@ repositories:
     type: git
     url: git@github.com:PR2/pr2_kinematics.git
     version: kinetic-devel
+  orocos_kinematics_dynamics:
+    type: git
+    url: https://github.com/orocos/orocos_kinematics_dynamics.git
+    version: v1.5.1
