+++ src/pycram/process_module.py
@@ -7,6 +7,7 @@ ProcessModule -- implementation of process modules.
 from __future__ import annotations
 import inspect
 import threading
+from threading import Lock
 import time
 from abc import ABC
 from typing_extensions import Callable, Type, Any, Union
@@ -61,7 +62,7 @@ class ProcessModule:
         with self._lock:
             ret = self._execute(designator)
             if ProcessModule.execution_delay:
-                time.sleep(0.5)
+                rospy.sleep(0.5)
 
         return ret
 
@@ -228,14 +229,20 @@ class ProcessModuleManager(ABC):
             return cls._instance
 
     def __init__(self, robot_name):
-        """
-        Registers the Process modules for this robot. The name of the robot has to match the name given in the robot
-        description.
-
-        :param robot_name: Name of the robot for which these Process Modules are intended
-        """
         self.robot_name = robot_name
         ProcessModuleManager.available_pms.append(self)
+        self._navigate_lock = Lock()
+        self._pick_up_lock = Lock()
+        self._place_lock = Lock()
+        self._looking_lock = Lock()
+        self._detecting_lock = Lock()
+        self._move_tcp_lock = Lock()
+        self._move_arm_joints_lock = Lock()
+        self._world_state_detecting_lock = Lock()
+        self._move_joints_lock = Lock()
+        self._move_gripper_lock = Lock()
+        self._open_lock = Lock()
+        self._close_lock = Lock()
 
     @staticmethod
     def get_manager() -> Union[ProcessModuleManager, None]:
