+++ src/pycram/process_modules/__init__.py
@@ -1,13 +1,13 @@
 from .pr2_process_modules import Pr2Manager
-from .boxy_process_modules import BoxyManager
-from .donbot_process_modules import DonbotManager
-from .hsrb_process_modules import HSRBManager
+#from .boxy_process_modules import BoxyManager
+#from .donbot_process_modules import DonbotManager
+#from .stretch_process_modules import StretchManager
+#from .hsrb_process_modules import HSRBManager
 from .default_process_modules import DefaultManager
-from .stretch_process_modules import StretchManager
 
 Pr2Manager()
-BoxyManager()
-DonbotManager()
-HSRBManager()
+#BoxyManager()
+#DonbotManager()
+#StretchManager()
+#HSRBManager()
 DefaultManager()
-StretchManager()
