+++ src/pycram/process_modules/default_process_modules.py
@@ -1,17 +1,20 @@
-from threading import Lock
-
 import numpy as np
+import rospy
+from typing_extensions import Any
 
 from ..datastructures.enums import JointType
 from ..external_interfaces.ik import request_ik
+from ..external_interfaces.robokudo import query
 from ..utils import _apply_ik
 from ..process_module import ProcessModule
+from ..external_interfaces import giskard
 from ..robot_description import RobotDescription
 from ..local_transformer import LocalTransformer
 from ..designators.motion_designator import *
 from ..world_reasoning import visible, link_pose_for_joint_config
 from ..world_concepts.world_object import Object
 from ..datastructures.world import World
+
 
 class DefaultNavigation(ProcessModule):
     """
@@ -21,6 +24,7 @@ class DefaultNavigation(ProcessModule):
     def _execute(self, desig: MoveMotion):
         robot = World.robot
         robot.set_pose(desig.target)
+
 
 class DefaultMoveHead(ProcessModule):
     """
@@ -34,8 +38,11 @@ class DefaultMoveHead(ProcessModule):
 
         local_transformer = LocalTransformer()
 
-        pan_link = RobotDescription.current_robot_description.kinematic_chains["neck"].links[0]
-        tilt_link = RobotDescription.current_robot_description.kinematic_chains["neck"].links[1]
+        # since its usually joint1, link1, joint2, link2, to be able to catch joint1, even though we only use the
+        # kinematic chain from link1 to link2, we need to use the link0 as the first link, even though its technically
+        # not in this kinematic chain
+        pan_link = RobotDescription.current_robot_description.kinematic_chains["neck"].links[1]
+        tilt_link = RobotDescription.current_robot_description.kinematic_chains["neck"].links[2]
 
         pan_joint = RobotDescription.current_robot_description.kinematic_chains["neck"].joints[0]
         tilt_joint = RobotDescription.current_robot_description.kinematic_chains["neck"].joints[1]
@@ -43,7 +50,8 @@ class DefaultMoveHead(ProcessModule):
         pose_in_tilt = local_transformer.transform_pose(target, robot.get_link_tf_frame(tilt_link))
 
         new_pan = np.arctan2(pose_in_pan.position.y, pose_in_pan.position.x)
-        new_tilt = np.arctan2(pose_in_tilt.position.z, pose_in_tilt.position.x ** 2 + pose_in_tilt.position.y ** 2) * -1
+        new_tilt = np.arctan2(pose_in_tilt.position.z,
+                              np.sqrt(pose_in_tilt.position.x ** 2 + pose_in_tilt.position.y ** 2)) * -1
 
         current_pan = robot.get_joint_position(pan_joint)
         current_tilt = robot.get_joint_position(tilt_joint)
@@ -62,7 +70,8 @@ class DefaultMoveGripper(ProcessModule):
         robot = World.robot
         gripper = desig.gripper
         motion = desig.motion
-        for joint, state in RobotDescription.current_robot_description.get_arm_chain(gripper).get_static_gripper_state(motion).items():
+        for joint, state in RobotDescription.current_robot_description.get_arm_chain(gripper).get_static_gripper_state(
+                motion).items():
             robot.set_joint_position(joint, state)
 
 
@@ -116,10 +125,15 @@ class DefaultMoveArmJoints(ProcessModule):
 
 
 class DefaultMoveJoints(ProcessModule):
+    """
+    Process Module for generic joint movements, is not confined to the arms but can move any joint of the robot
+    """
+
     def _execute(self, desig: MoveJointsMotion):
         robot = World.robot
         for joint, pose in zip(desig.names, desig.positions):
             robot.set_joint_position(joint, pose)
+        return ",", ","
 
 
 class DefaultWorldStateDetecting(ProcessModule):
@@ -148,14 +162,15 @@ class DefaultOpen(ProcessModule):
         _move_arm_tcp(goal_pose, World.robot, desig.arm)
 
         desig.object_part.world_object.set_joint_position(container_joint,
-                                                              part_of_object.get_joint_limits(
-                                                                  container_joint)[1])
+                                                          part_of_object.get_joint_limits(
+                                                              container_joint)[1])
 
 
 class DefaultClose(ProcessModule):
     """
     Low-level implementation that lets the robot close a grasped container, in simulation
     """
+
     def _execute(self, desig: ClosingMotion):
         part_of_object = desig.object_part.world_object
 
@@ -167,8 +182,8 @@ class DefaultClose(ProcessModule):
         _move_arm_tcp(goal_pose, World.robot, desig.arm)
 
         desig.object_part.world_object.set_joint_position(container_joint,
-                                                              part_of_object.get_joint_limits(
-                                                                  container_joint)[0])
+                                                          part_of_object.get_joint_limits(
+                                                              container_joint)[0])
 
 
 def _move_arm_tcp(target: Pose, robot: Object, arm: Arms) -> None:
@@ -180,57 +195,213 @@ def _move_arm_tcp(target: Pose, robot: Object, arm: Arms) -> None:
     _apply_ik(robot, inv)
 
 
+###########################################################
+######### Process Modules for the Real Robot ##############
+###########################################################
+
+class DefaultNavigationReal(ProcessModule):
+    """
+    Process module for the real robot that sends a cartesian goal to giskard to move the robot base
+    """
+
+    def _execute(self, designator: MoveMotion) -> Any:
+        rospy.logdebug(f"Sending goal to giskard to Move the robot")
+        giskard.achieve_cartesian_goal(designator.target, RobotDescription.current_robot_description.base_link, "map")
+
+
+class DefaultMoveHeadReal(ProcessModule):
+    """
+    Process module for the real robot to move that such that it looks at the given position. Uses the same calculation
+    as the simulated one
+    """
+
+    def _execute(self, desig: LookingMotion):
+        target = desig.target
+        robot = World.robot
+
+        local_transformer = LocalTransformer()
+        pan_link = RobotDescription.current_robot_description.kinematic_chains["neck"].links[0]
+        tilt_link = RobotDescription.current_robot_description.kinematic_chains["neck"].links[1]
+
+        pan_joint = RobotDescription.current_robot_description.kinematic_chains["neck"].joints[0]
+        tilt_joint = RobotDescription.current_robot_description.kinematic_chains["neck"].joints[1]
+        pose_in_pan = local_transformer.transform_pose(target, robot.get_link_tf_frame(pan_link))
+        pose_in_tilt = local_transformer.transform_pose(target, robot.get_link_tf_frame(tilt_link))
+
+        new_pan = np.arctan2(pose_in_pan.position.y, pose_in_pan.position.x)
+        new_tilt = np.arctan2(pose_in_tilt.position.z,
+                              np.sqrt(pose_in_tilt.position.x ** 2 + pose_in_tilt.position.y ** 2)) * -1
+
+        current_pan = robot.get_joint_position(pan_joint)
+        current_tilt = robot.get_joint_position(tilt_joint)
+
+        robot.set_joint_position(pan_joint, new_pan + current_pan)
+        robot.set_joint_position(tilt_joint, new_tilt + current_tilt)
+
+
+class DefaultDetectingReal(ProcessModule):
+    """
+    Process Module for the real robot that tries to detect an object fitting the given object description. Uses Robokudo
+    for perception of the environment.
+    """
+
+    def _execute(self, designator: DetectingMotion) -> Any:
+        query_result = query(ObjectDesignatorDescription(types=[designator.object_type]))
+        # print(query_result)
+        obj_pose = query_result["ClusterPoseBBAnnotator"]
+
+        lt = LocalTransformer()
+        obj_pose = lt.transform_pose(obj_pose, World.robot.get_link_tf_frame("torso_lift_link"))
+        obj_pose.orientation = [0, 0, 0, 1]
+        obj_pose.position.x += 0.05
+
+        bullet_obj = World.current_world.get_objects_by_type(designator.object_type)
+        if bullet_obj:
+            bullet_obj[0].set_pose(obj_pose)
+            return bullet_obj[0]
+        elif designator.object_type == ObjectType.JEROEN_CUP:
+            cup = Object("cup", ObjectType.JEROEN_CUP, "jeroen_cup.stl", pose=obj_pose)
+            return cup
+        elif designator.object_type == ObjectType.BOWL:
+            bowl = Object("bowl", ObjectType.BOWL, "bowl.stl", pose=obj_pose)
+            return bowl
+
+        return bullet_obj[0]
+
+
+class DefaultMoveTCPReal(ProcessModule):
+    """
+    Moves the tool center point of the real robot while avoiding all collisions
+    """
+
+    def _execute(self, designator: MoveTCPMotion) -> Any:
+        lt = LocalTransformer()
+        pose_in_map = lt.transform_pose(designator.target, "map")
+
+        if designator.allow_gripper_collision:
+            giskard.allow_gripper_collision(designator.arm.name)
+        giskard.achieve_cartesian_goal(pose_in_map,
+                                       RobotDescription.current_robot_description.get_arm_chain(
+                                           designator.arm).get_tool_frame(),
+                                       RobotDescription.current_robot_description.base_link)
+
+
+class DefaultMoveArmJointsReal(ProcessModule):
+    """
+    Moves the arm joints of the real robot to the given configuration while avoiding all collisions
+    """
+
+    def _execute(self, designator: MoveArmJointsMotion) -> Any:
+        joint_goals = {}
+        if designator.left_arm_poses:
+            joint_goals.update(designator.left_arm_poses)
+        if designator.right_arm_poses:
+            joint_goals.update(designator.right_arm_poses)
+        giskard.avoid_all_collisions()
+        giskard.achieve_joint_goal(joint_goals)
+
+
+class DefaultMoveJointsReal(ProcessModule):
+    """
+    Moves any joint using giskard, avoids all collisions while doint this.
+    """
+
+    def _execute(self, designator: MoveJointsMotion) -> Any:
+        name_to_position = dict(zip(designator.names, designator.positions))
+        giskard.avoid_all_collisions()
+        giskard.achieve_joint_goal(name_to_position)
+
+
+class DefaultMoveGripperReal(ProcessModule):
+    """
+    Opens or closes the gripper of the real robot, gripper uses an action server for this instead of giskard
+    """
+
+    def _execute(self, designator: MoveGripperMotion) -> Any:
+        raise NotImplementedError(f"There is DefaultMoveGripperReal process module")
+
+
+class DefaultOpenReal(ProcessModule):
+    """
+    Tries to open an already grasped container
+    """
+
+    def _execute(self, designator: OpeningMotion) -> Any:
+        giskard.achieve_open_container_goal(
+            RobotDescription.current_robot_description.get_arm_chain(designator.arm).get_tool_frame(),
+            designator.object_part.name)
+
+
+class DefaultCloseReal(ProcessModule):
+    """
+    Tries to close an already grasped container
+    """
+
+    def _execute(self, designator: ClosingMotion) -> Any:
+        giskard.achieve_close_container_goal(
+            RobotDescription.current_robot_description.get_arm_chain(designator.arm).get_tool_frame(),
+            designator.object_part.name)
+
+
 class DefaultManager(ProcessModuleManager):
 
     def __init__(self):
         super().__init__("default")
-        self._navigate_lock = Lock()
-        self._looking_lock = Lock()
-        self._detecting_lock = Lock()
-        self._move_tcp_lock = Lock()
-        self._move_arm_joints_lock = Lock()
-        self._world_state_detecting_lock = Lock()
-        self._move_joints_lock = Lock()
-        self._move_gripper_lock = Lock()
-        self._open_lock = Lock()
-        self._close_lock = Lock()
 
     def navigate(self):
         if ProcessModuleManager.execution_type == "simulated":
             return DefaultNavigation(self._navigate_lock)
+        elif ProcessModuleManager.execution_type == "real":
+            return DefaultNavigationReal(self._navigate_lock)
 
     def looking(self):
         if ProcessModuleManager.execution_type == "simulated":
             return DefaultMoveHead(self._looking_lock)
+        elif ProcessModuleManager.execution_type == "real":
+            return DefaultMoveHeadReal(self._looking_lock)
 
     def detecting(self):
         if ProcessModuleManager.execution_type == "simulated":
             return DefaultDetecting(self._detecting_lock)
+        elif ProcessModuleManager.execution_type == "real":
+            return DefaultDetectingReal(self._detecting_lock)
 
     def move_tcp(self):
         if ProcessModuleManager.execution_type == "simulated":
             return DefaultMoveTCP(self._move_tcp_lock)
+        elif ProcessModuleManager.execution_type == "real":
+            return DefaultMoveTCPReal(self._move_tcp_lock)
 
     def move_arm_joints(self):
         if ProcessModuleManager.execution_type == "simulated":
             return DefaultMoveArmJoints(self._move_arm_joints_lock)
+        elif ProcessModuleManager.execution_type == "real":
+            return DefaultMoveArmJointsReal(self._move_arm_joints_lock)
 
     def world_state_detecting(self):
-        if ProcessModuleManager.execution_type == "simulated":
+        if ProcessModuleManager.execution_type == "simulated" or ProcessModuleManager.execution_type == "real":
             return DefaultWorldStateDetecting(self._world_state_detecting_lock)
 
     def move_joints(self):
         if ProcessModuleManager.execution_type == "simulated":
             return DefaultMoveJoints(self._move_joints_lock)
+        elif ProcessModuleManager.execution_type == "real":
+            return DefaultMoveJointsReal(self._move_joints_lock)
 
     def move_gripper(self):
         if ProcessModuleManager.execution_type == "simulated":
             return DefaultMoveGripper(self._move_gripper_lock)
+        elif ProcessModuleManager.execution_type == "real":
+            return DefaultMoveGripperReal(self._move_gripper_lock)
 
     def open(self):
         if ProcessModuleManager.execution_type == "simulated":
             return DefaultOpen(self._open_lock)
+        elif ProcessModuleManager.execution_type == "real":
+            return DefaultOpenReal(self._open_lock)
 
     def close(self):
         if ProcessModuleManager.execution_type == "simulated":
             return DefaultClose(self._close_lock)
+        elif ProcessModuleManager.execution_type == "real":
+            return DefaultCloseReal(self._close_lock)
